+++ src/ofxGgmlSam.h
@@ -2,6 +2,7 @@
 
 #include "ofxGgml.h"
 #include "ofxGgmlSamVersion.h"
+#include "ofxGgmlSam/ofxGgmlSamExternalBackend.h"
 #include "ofxGgmlSam/ofxGgmlSamInference.h"
 #include "ofxGgmlSam/ofxGgmlSamTypes.h"
 #include "ofxGgmlSam/ofxGgmlSamUtils.h"